Since there were some unsmoothnesses in the flow of the game, that had to do with framerates and distorted framerates I corrected the flaw.  Apart from that there is now a function called Activate_Conservative_Frame_Computation(); that should be called whenever the game flow is stopped for possibly longer periods of time, e.g. in elevators, consoles, takeover, options menu, debug menu.  The new code seems to work pretty well, at least on my machine.

+++ src/struct.h
@@ -129,7 +129,7 @@ typedef struct
   finepoint pos;		/* current position in level levelnum */
   float health;			/* the max. possible energy in the moment */
   float energy;			/* current energy */
-  byte firewait;		/* counter after fire */
+  float firewait;		/* counter after fire */
   float phase;			/* the current phase of animation */
   int autofire;			/* Status of the Firecontrolautomatics */
   int vneut;			/* Status of Velocityneutralizer for the gun */
